move artists to tracing package
