+++ src/main/java/sc/fiji/snt/tracing/artist/BiSearchArtist.java
@@ -20,8 +20,9 @@
  * #L%
  */
 
-package sc.fiji.snt;
+package sc.fiji.snt.tracing.artist;
 
+import sc.fiji.snt.TracerCanvas;
 import sc.fiji.snt.hyperpanes.MultiDThreePanes;
 import sc.fiji.snt.tracing.BiSearch;
 import sc.fiji.snt.tracing.BiSearchNode;
