+++ src/main/java/sc/fiji/snt/tracing/artist/SearchThreadArtist.java
@@ -20,8 +20,9 @@
  * #L%
  */
 
-package sc.fiji.snt;
+package sc.fiji.snt.tracing.artist;
 
+import sc.fiji.snt.TracerCanvas;
 import sc.fiji.snt.hyperpanes.MultiDThreePanes;
 import sc.fiji.snt.tracing.SearchInterface;
 import sc.fiji.snt.tracing.SearchNode;
